SSL-07 CS final

+++ index.js
@@ -82,10 +82,10 @@ async function run() {
       const order = req.body;
       // const result = await orderCollection.insertOne(order);
       // res.send(result);
-      // const { service, email, address } = order;
-      // if (!service || !email || !address) {
-      //   return res.send({ error: 'Please provide all the information' });
-      // }
+      const { service, email, address } = order;
+      if (!service || !email || !address) {
+        return res.send({ error: 'Please provide all the information' });
+      }
       const orderedService = await serviceCollection.findOne({_id: ObjectId(order.service), });
       console.log(orderedService);
       //  res.send(orderedService);
@@ -94,9 +94,9 @@ async function run() {
         total_amount: orderedService.price,
         currency: order.currency,
         tran_id: transactionId, // use unique tran_id for each api call
-        success_url: `http://localhost:5000/payment/success?transactionId=${transactionId}`,
-        fail_url: `http://localhost:5000/payment/fail?transactionId=${transactionId}`,
-        cancel_url: 'http://localhost:5000/payment/cancel',
+        success_url: `${process.env.SERVER_URL}/payment/success?transactionId=${transactionId}`,
+        fail_url: `${process.env.SERVER_URL}/payment/fail?transactionId=${transactionId}`,
+        cancel_url: `${process.env.SERVER_URL}/payment/cancel`,
         ipn_url: 'http://localhost:3030/ipn',
         shipping_method: 'Courier',
         product_name: 'Computer.',
@@ -137,33 +137,30 @@ async function run() {
     });
   });
 
-  app.post('/payment/success', async (req, res) => {
-    // console.log('Success')
-    const { transactionId } = req.query;
-    // console.log(transactionId)
-    // if (!transactionId) {
-    //   return res.redirect(`${process.env.CLIENT_URL}/payment/fail`);
-    // }
+  app.post('/payment/success', async (req, res) => {  // console.log('Success')
+    const { transactionId } = req.query; // console.log(transactionId)
+    if (!transactionId) {
+      return res.redirect(`${process.env.CLIENT_URL}/payment/fail`);
+    }
     const result = await orderCollection.updateOne(
       { transactionId },
       { $set: { paid: true, paidAt: new Date() } }
     );
     if (result.modifiedCount > 0) {
-      res.redirect(`http://localhost:3000/payment/success?transactionId=${transactionId}`
-        // `${process.env.CLIENT_URL}/payment/success?transactionId=${transactionId}`
+      res.redirect(
+        `${process.env.CLIENT_URL}/payment/success?transactionId=${transactionId}`
       );
     }
   });
 
   app.post('/payment/fail', async (req, res) => {
     const { transactionId } = req.query;
-    // if (!transactionId) {
-    //   return res.redirect(`${process.env.CLIENT_URL}/payment/fail`);
-    // }
+    if (!transactionId) {
+      return res.redirect(`${process.env.CLIENT_URL}/payment/fail`);
+    }
     const result = await orderCollection.deleteOne({ transactionId });
     if (result.deletedCount) {
-      res.redirect('http://localhost:3000/payment/fail');
-      // res.redirect(`${process.env.CLIENT_URL}/payment/fail`);
+      res.redirect(`${process.env.CLIENT_URL}/payment/fail`);
     }
   });
